pyproject.toml
[tool.poetry]
name = "mumoco"
version = "0.2.3"
description = "This is tool helps to work with multiple conan modules simultaneously."
license = "MIT"
homepage = "https://github.com/disroop/mumoco"
authors = ["michel <michel.meyer@disroop.ch>"]

[tool.poetry.urls]
Issues = "https://github.com/disroop/mumoco/issues"

[tool.poetry.dependencies]
python = "^3.8"
conan = "^1.36.0"
deserialize = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
flake8 = "^3.9.2"
pytest-cov = "^2.12.1"
mypy = "^0.902"
black = "21.6b0"
isort = "^5.9.1"
bandit = "^1.7.0"
invoke = "^1.5.0"
pylint = "^2.8.2"

[tool.poetry.scripts]
mumoco = "src.mumoco_cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
enable = ''',
    unused-import,
    unused-wildcard-import,
'''
disable = ''',
    fixme,
    missing-module-docstring,
    missing-class-docstring,
    missing-function-docstring,
 '''

[tool.pylint.format]
max-line-length = 120

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

# [tool.flake8]
# Flake8 currently does not support pyproject.toml.
# See .flake8 instead.

[tool.mypy]
files = "src/**/*.py"
allow_untyped_decorators = false
warn_unused_configs = true
allow_subclassing_any = false
allow_untyped_calls = false
allow_untyped_defs = false
allow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
ignore_missing_imports = false
pretty = true

[[tool.mypy.overrides]]
module = [
    "deserialize",
    "conans.client.conan_api",
    "conans.errors"
]
ignore_missing_imports = true
# Mypi  currently does not support pyproject.toml.
# See mypy.ini instead.